[NVIDIA] Replace proxy fences with buffer-region analysis
